pyproject.toml
[tool.poetry]
name = "BauerDude"
version = "0.1.0"
description = ""
authors = ["ThaRising <mail@domain.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
django = "3.1.1"
djangorestframework = "3.12.1"
drf-yasg = "1.20.0"
Pillow = "^8.0.1"
django-debug-toolbar = "^2.2"
djangorestframework-simplejwt = "^4.4.0"
django-cors-headers = "^3.5.0"
drizm-django-commons = "^0.3.0"
drizm-commons = "^0.5.0"
psycopg2-binary = "^2.8.6"
dj-database-url = "^0.5.0"
whitenoise = {extras = ["brotli"], version = "^5.2.0"}
pandas = "^1.2.1"
openpyxl = "^3.0.6"
django-silk = "^4.1.0"

[tool.poetry.dev-dependencies]
isort = "^5.1.4"
autoflake = "^1.3.1"
black = "^19.10b0"

[tool.black]
line-length = 80
target_version = ["py38"]
include = '\.pyi?$'

[tool.isort]
known_first_party = ["BauerDude", "tests"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
